+++ src/test/java/seedu/address/logic/commands/deckcommands/AddDeckCommandTest.java
@@ -8,28 +8,15 @@ import static seedu.address.testutil.Assert.assertThrows;
 import static seedu.address.testutil.TypicalDecks.VALID_DECK_HISTORY;
 import static seedu.address.testutil.TypicalDecks.VALID_DECK_SCIENCE;
 
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Optional;
-import java.util.function.Predicate;
 
 import org.junit.jupiter.api.Test;
 
-import javafx.collections.ObservableList;
-import javafx.util.Pair;
-import seedu.address.commons.core.GuiSettings;
-import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.commandresult.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Model;
-import seedu.address.model.ModelState;
-import seedu.address.model.ReadOnlyMasterDeck;
-import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.card.Card;
+import seedu.address.model.ModelStub;
 import seedu.address.model.deck.Deck;
-import seedu.address.model.review.Review;
-import seedu.address.model.tag.Tag;
 
 public class AddDeckCommandTest {
 
@@ -82,210 +69,9 @@ public class AddDeckCommandTest {
     }
 
     /**
-     * A default model stub that have all the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getMasterDeckFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setMasterDeckFilePath(Path deckFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addCard(Card card) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setMasterDeck(ReadOnlyMasterDeck newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyMasterDeck getMasterDeck() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasCard(Card card) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteCard(Card target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setCard(Card target, Card editedCard) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Card> getFilteredCardList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Deck> getFilteredDeckList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredCardList(Predicate<Card> predicate) {
-            return; // AddCommand does call updateFilteredCardList method
-        }
-
-        @Override
-        public void updateFilteredDeckList(Predicate<Deck> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        /* NEWLY ADDED COMMANDS TO SUPPORT DECK LIST */
-        @Override
-        public Optional<Deck> getSelectedDeck() {
-            return Optional.of(new Deck("Default"));
-        }
-
-        @Override
-        public void addDeck(Deck deck) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasDeck(Deck deck) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setDeck(Deck target, Deck editedDeck) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void moveCards(Deck target, Deck editedDeck) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteDeck(Deck key) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void selectDeck(Index deckIndex) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void unselectDeck() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public String getSelectedDeckName() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public int getDeckSize(int deckIndex) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        public Optional<Review> getReview() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void reviewDeck(Index deckIndex) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setNumCardsPerReview(int reviewLimit) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Pair<String, String> > getReviewStatsList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void endReview() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public String getReviewDeckName() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Pair<String, String> > getReviewDeckNameList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-
-        @Override
-        public void flipCard() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean goToPrevCard() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean goToNextCard() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ModelState getState() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void tagCurrentCardInReview(Tag tag) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean isReviewCardFlipped() {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
      * A Model stub that contains a single deck.
      */
-    private class ModelStubWithDeck extends ModelStub {
+    private static class ModelStubWithDeck extends ModelStub {
         private final Deck deck;
 
         ModelStubWithDeck(Deck deck) {
@@ -303,7 +89,7 @@ public class AddDeckCommandTest {
     /**
      * A Model stub that always accept the deck being added.
      */
-    private class ModelStubAcceptingDeckAdded extends ModelStub {
+    private static class ModelStubAcceptingDeckAdded extends ModelStub {
         final ArrayList<Deck> decksAdded = new ArrayList<>();
 
         @Override
@@ -311,6 +97,7 @@ public class AddDeckCommandTest {
             requireNonNull(deck);
             return decksAdded.stream().anyMatch(deck::isSameDeck);
         }
+
         @Override
         public void addDeck(Deck deck) {
             requireNonNull(deck);
